Win/Lose
Win/Lose is now fully integrated with timer ending and goal object being picked up. playerController and mouseMove now disables when the game is in either win or lose, with the mouse pointer beomcing visible so player can click through any UI elements that we implement for game win/lose.

State Engine includes a tutorial state where we can have other UI elements appear at the beginning of the game after start button is pressed but before the timer starts, we would need to code that though before it operates in that way.